fixup test case: not convertible to target
