+++ internal/e2e/core/cluster_node.go
@@ -1,7 +1,6 @@
 package core
 
 import "github.com/hashicorp/mog/internal/e2e/core/inner"
-import "github.com/hashicorp/mog/internal/e2e/core/status"
 
 type Label string
 
@@ -13,7 +12,7 @@ type ClusterNode struct {
 
 	O *Other
 	I inner.Inner
-	S status.Status[status.NodeCondition] // for testing generic structs
+	S Status[NodeCondition] // for testing generic structs
 
 	F1 Workload  // for testing struct-to-struct
 	F2 *Workload // for testing ptr-to-ptr
@@ -64,3 +63,18 @@ type Workload struct {
 type Other struct {
 	N int
 }
+
+type Status[T ConditionType] struct {
+	Conditions []Condition[T]
+}
+
+type Condition[T ConditionType] struct {
+	Type T
+}
+
+type ConditionType interface {
+	NodeCondition | WorkloadCondition
+}
+
+type NodeCondition string
+type WorkloadCondition string
